#1 Change ---------- Track class it's more object oriented.
Files
-----
Camera.cxx
History.cxx
start_tuxkart.cxx
TrackManager.cxx
World.cxx
WorldScreen.cxx
Track.h


#2 Changes
----------
To the function spatialToTrack, it finds a the closest point in the driveline that is at most two positions away in the list from the hint given, for example, hint 85 would search for the closest point in the range of [83,87].
Some minor speed improvements in functions which use ask for the size() of a vector in a loop.
Modified the driveline loading to load two files, which indicate the left and right borders, which are used to find the center driveline and the track width.

File
-----
Track.cxx


#3 Change
----------
Added comments.

Files
-----
CollectableManager.cxx
Kart.h


#4 Change
----------
Clear attachments when reseting the track and eliminated a quick fix that is not needed anymore.

File
-----
Kart.cxx


#5 Change
----------
Added abs_velocity, which gives the velocity of the kart in respect to the world.

Files
-----
Moveable.cxx
Moveable.h

#6 Change
----------
Improved AI, still very far from finished.

Files
-----
AutoKart.cxx
Autokart.h



git-svn-id: https://svn.code.sf.net/p/supertuxkart/code@575 178a84e3-b1eb-0310-8ba1-8eac791a3b58