Removed the warning " [StringUtils] WARNING: Invalid HTML entity in 'Berny & STKRudy85' (we only support &#...; type HTML entities)"; #364.
The reason for this warning message is that irrXML already decodes "&amp;"
(in attribute values) - so even though the '&' is encoded correctly in
the xml as '&amp;', STK only receives a '&' from irrlicht, which then triggers the
warning message. On the other hand, irrXML does not decode &#...; hex
values. 

So short of changing our encoding of "&" I don't see another choice
except just removing this warning.


git-svn-id: https://svn.code.sf.net/p/supertuxkart/code@9510 178a84e3-b1eb-0310-8ba1-8eac791a3b58