GP start order option to user's config + add option to have players always last
In light of #28, the GP start order option is probably more at home in a user's
configuration. While at it, also add an option to keep the player always last.
This is more of a challenge for the player than just reverse mode since your
direct opponents _will_ be at the forefront while you have to play catch up.


git-svn-id: https://svn.code.sf.net/p/supertuxkart/code@11017 178a84e3-b1eb-0310-8ba1-8eac791a3b58