Correct author and license information for the "Boom boom boom" song.
The real author of this song is Keith Baylis (aka Vim). The song has been composed in 1993 and was available in Tuxkart/Supertuxkart since 2002.

After direct contact with the real author, he agreed that this song can be licensed under CC-BY-SA v4.0 license.

More info in this topic:
http://forum.freegamedev.net/viewtopic.php?f=17&t=6562

Also note that source files (the mod file and lossless wav file) are available in our media repo:
https://sourceforge.net/p/supertuxkart/code/HEAD/tree/media/trunk/music/mods/

It solves the bugs:
https://github.com/supertuxkart/stk-code/issues/2577
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=832062

git-svn-id: https://svn.code.sf.net/p/supertuxkart/code@16762 178a84e3-b1eb-0310-8ba1-8eac791a3b58